implement SegmentMatchTrafficAlgorithm
