+++ shardingsphere-kernel/shardingsphere-traffic/shardingsphere-traffic-api/src/main/java/org/apache/shardingsphere/traffic/api/config/TrafficRuleConfiguration.java
@@ -20,7 +20,6 @@ package org.apache.shardingsphere.traffic.api.config;
 import lombok.Getter;
 import org.apache.shardingsphere.infra.config.algorithm.ShardingSphereAlgorithmConfiguration;
 import org.apache.shardingsphere.infra.config.scope.GlobalRuleConfiguration;
-import org.apache.shardingsphere.traffic.api.config.strategy.TrafficStrategyConfiguration;
 
 import java.util.Collection;
 import java.util.LinkedHashMap;
